requirements.txt
langchain
langchain-groq
langchain_community
python-dotenv
gradio